moved gather routines into the already existing 'read' utility function
